+++ Assets/Tests/PostEffects/InverColorEffect.cs
@@ -1,0 +1,11 @@
+﻿using UnityEngine;
+
+public class InverColorEffect : MonoBehaviour {
+
+	public Material material;
+
+	private void OnRenderImage(RenderTexture source, RenderTexture destination)
+	{
+		Graphics.Blit(source, destination, material);
+	}
+}